refactor: Move shared functions to utils
